+++ lib/view_component/template.rb
@@ -59,6 +59,14 @@ module ViewComponent
       @component.define_method(safe_method_name, @component.instance_method(@call_method_name))
     end
 
+    def safe_method_name_call
+      return safe_method_name unless inline_call?
+
+      "maybe_escape_html(#{safe_method_name}) " \
+      "{ Kernel.warn('WARNING: The #{@component} component rendered HTML-unsafe output. " \
+      "The output will be automatically escaped, but you may want to investigate.') } "
+    end
+
     def requires_compiled_superclass?
       inline_call? && !defined_on_self?
     end