+++ test/sandbox/app/models/coupon.rb
@@ -1,0 +1,1 @@
+Coupon = Struct.new(:percent_off, keyword_init: true)